ready to test
First complete draft of new make4FGLxml.py script, rename some files, need to test and then integrate with GUI functions
