+++ make4FGLxml.py
@@ -1,99 +1,503 @@
 #!/usr/bin/env python
 
-class srcList:
+from build_model.utilities import (
+        angular_separation,
+        get_ROI_from_event_file,
+        build_region)
+
+from build_model.model_components import Spectrum,Spatial
+
+import os,warnings
+import numpy as np
+import pandas as pd
+from functools import reduce
+import astropy.io.fits as pyfits
+from xml.dom import minidom
+
+class SourceList:
 	#arguments are:
-	#sources (string, filename of LAT source list fits file in catalog format)
-	#ft1 (string, filename of event file for which the xml will be used, only used to extract ROI info)
-	#out (string, name of output xml file, defaults to mymodel.xml)
-	def __init__(self,sources,ft1,out='mymodel.xml',DRversion=3):
-		if not fileCheck(sources): #check that file exists
-			print("Error:  %s not found."%sources)
-			return
-		if fileCheck(out):
-			print('Warning: %s already exists, file will be overwritten if you proceed with makeModel.'%out)
-		if DRversion not in [1,2,3]:
-			print('Invalid choice, %i, of data release version.  Must be 1, 2, or 3.\nWill produce file assuming DR3')
-			self.DR=3
-			self.VarLim=24.725
-		else:
-			self.DR=DRversion
-			self.VarLim=(24.725 if DRversion==3 else 21.666 if DRversion==2 else 18.475)
-		self.srcs=sources
-		self.out=out
-		self.roi=getPos(ft1)
-	
-	#define a quick print function to make sure everything looks irght
+	#catalog_file (string, filename of LAT source list fits file in catalog format)
+	#ROI (either a list with [RA,DEC,radius] or string with name of event file to extract ROI info from)
+	#output_name (string, name of output xml file, defaults to my_model.xml)
+        #DR (int, data release version, default is 3)
+        def __init__(self,catalog_file,ROI,output_name='my_model.xml',DR=3,write_directory=''):
+                #some sanity checks on input values
+                if not os.path.exists(catalog_file):
+                        raise FileNotFoundError(2,'Could not access catalog file',catalog_file)
+                
+		if os.path.exists(output_name):
+                        warnings.warn(f'Region model {output_name} already exists, will be overwritten.')
+
+                if DR not in [1,2,3]:
+                        raise ValueError(f'DR={DR} is an invalid choice of data release, must be 1, 2, or 3.')
+
+                self.catalog_file=catalog_file
+                self.output_name=os.path.join(write_directory,output_name)
+                self.write_directory=write_directory
+                self.DR=DR
+                self.variability_threshold=24.725 if DR==3 else 21.666 if DR==1 else 18.475
+
+                if isinstance(ROI,list):
+                        self.ROI=ROI
+                        
+                elif os.path.exists(ROI)
+                        self.ROI=get_ROI_from_event_file(ROI)
+                else:
+                        raise FileNotFoundError(2,'Could not access event file',ROI)
+        
+	#define a quick print function to make sure everything looks right
 	def Print(self):
-		print('Source list file: ',self.srcs)
-		print('Output file name: ',self.out)
-		print('Selecting %s degrees around (ra,dec)=(%s,%s)'%(self.roi[2],self.ra,self.dec))
+		print('Catalog file: ',self.catalog_file)
+		print('Output file name: ',self.output_name)
+		print(f'Selecting {self.roi[2]:.1f} degrees around (ra,dec)=({self.roi[0]:.3f},{self.roi[1]:.3f})')
 	
 	#make the xml file
 	#arguments are:
-	#GDfile (str) -- optional, location and name of Galactic diffuse model to use
-	#GDname (str) -- optional, name of Galactic diffuse component to use in xml model
-	#ISOfile (str) -- optional, location and name of Isotropic diffuse template to use
-	#ISOname (str) -- optional, name of Isotropic diffuse component to use in xml model
-	#normsOnly (bool) -- optional, flag to only set normalizations parameters free
-	#extDir (str) -- optional, directory with extended source templates
-	#radLim (float) -- optional, radius in degrees from center of ROI beyond which source parameters are fixed
-	#maxRad (float) -- optional, absolute maximum radius beyond which sources are fixed, this may be necessary when doing binned analysis and a variable source beyond radLim would be set free but this source is beyond the boundaries of the square region used for the binned likelihood
-	#ExtraRad (float) -- optional, radius beyond ROI radius in event file out to which sources will be included with fixed parameters, defaul tof 10 is good for analyses starting around 100 MeV, but for higher energy fits this can be decreased
-	#sigFree (float) -- optional, average significance, using FITS catalog file, below which source parameters are fixed, even if within radLim.  This corresponds to TS_value if using the XML catalog file.
-	#varFree (float) -- optional, variability index above which source parameters are free, if beyond radLim and/or below sigFree only the normalization parameter is set free
-	#psForce (bool) -- optional, flag to force extended sources to be point sources
-	#E2CAT (bool) -- optional, flag to force use catalog names for extended sources (only matters if using catalog FITS file)
-	#makeRegion (bool) -- optional, flag to also generate ds9 region file
-	#GIndexFree (bool) -- optional, the Galactic diffuse is given a power-law spectral shape but the by default the index is frozen, setting this flag to True allows that to be free for additional freedom in diffuse fit
-	#oldNames (bool) -- optional, flag to use the naming convention from make1FGLxml.py and make2FGLxml.py with a leading underscore and no spaces
-	def makeModel(self,GDfile="$(FERMI_DIR)/refdata/fermi/galdiffuse/gll_iem_v07.fits",GDname='gll_iem_v07',
-                      ISOfile="$(FERMI_DIR)/refdata/fermi/galdiffuse/iso_P8R3_SOURCE_V3_v1.txt",ISOname='iso_P8R3_SOURCE_V3_v1',
-                      normsOnly=False,extDir='',radLim=-1,maxRad=None,ExtraRad=10,sigFree=5,varFree=True,psForce=False,E2CAT=False,
-                      makeRegion=True,GIndexFree=False,wd='',oldNames=False):
-                
-		self.radLim=(self.roi[2] if radLim<=0 else radLim)
-		self.maxRad=(self.radLim if maxRad==None else maxRad)
-		if self.maxRad<self.radLim:
-			print("NOTE: maxRad (%.1f deg) is less than radLim (%.1f deg), meaning maxRad parameter is useless"%(self.maxRad,self.radLim))
-		self.var=varFree
-		self.psF=psForce
-		self.E2C=E2CAT
-		self.nO=normsOnly
-		extDir=(extDir if extDir!='' else '$(FERMI_DIR)/data/pyBurstAnalysisGUI/templates')#make sure the default for FSSC STs is correct
-		self.extD=(extDir if extDir[-1]=='/' else extDir+'/')
-		self.ER=ExtraRad
-		self.sig=sigFree
-		self.reg=makeRegion
-		self.GIF=GIndexFree
-		if makeRegion:
-			rhold=self.out.split('.')[:-1]
-			rhold=rhold[0].split('/')[-1]#just in case the user has specified the full path for the output XML model file
-			wd=(os.getcwd() if wd=='' else wd)
-			self.regFile=wd+'/ROI_'
-			for r in rhold:
-				self.regFile+=r
-			self.regFile+='.reg'
-		print('Creating file and adding sources from 4FGL')
-		#want ability to use either the FITS or xml versions of the catalog
-		#need to tweak FITS version to have new functionality and then work out xml version
-		if self.srcs.split('.')[-1]=='xml':
-			addSrcsXML(self,GDfile,GDname,ISOfile,ISOname,oldNames)
-		else:
-			addSrcsFITS(self,GDfile,GDname,ISOfile,ISOname,oldNames)
-	
-import astropy.io.fits as pyfits
-import os
-from xml.dom import minidom
-from xml.dom.minidom import parseString as pS
-from numpy import floor,log10,cos,sin,arccos,pi,array,log,exp
-acos=arccos
-print("This is make4FGLDR3xml version 01r09.")
-print("The default diffuse model files and names are for P8R3, 4FGL DRs, and the default fermitools directory locations.")
-d2r=pi/180.
-####varValue=24.725#21.666 for dr2,18.475 for 4fgl 
+	#galactic_file (str) -- optional, location and name of Galactic diffuse model to use
+	#galactic_name (str) -- optional, name of Galactic diffuse component to use in xml model
+	#isotropic_file (str) -- optional, location and name of Isotropic diffuse template to use
+	#isotropic_name (str) -- optional, name of Isotropic diffuse component to use in xml model
+	#norms_free_only (bool) -- optional, flag to only set normalizations parameters free
+	#extended_directory (str) -- optional, directory with extended source templates
+	#free_radius (float) -- optional, radius in degrees from center of ROI beyond which source parameters are fixed
+	#max_free_radius (float) -- optional, absolute maximum radius beyond which sources are fixed, this may be necessary when doing binned analysis and a variable source beyond free_radius would be set free but this source is beyond the boundaries of the square region used for the binned likelihood
+	#extra_radius (float) -- optional, radius beyond ROI radius in event file out to which sources will be included with fixed parameters, defaul tof 10 is good for analyses starting around 100 MeV, but for higher energy fits this can be decreased
+	#sigma_to_free (float) -- optional, average significance, using FITS catalog file, below which source parameters are fixed, even if within free_radius.  This corresponds to TS_value if using the XML catalog file.
+	#variable_free (float) -- optional, variability index above which source parameters are free, if beyond free_radius and/or below sigma_to_free only the normalization parameter is set free
+	#force_point_sources (bool) -- optional, flag to force extended sources to be point sources
+	#extended_catalog_names (bool) -- optional, flag to force use catalog names for extended sources (only matters if using catalog FITS file)
+	#make_region (bool) -- optional, flag to also generate ds9 region file
+	#galactic_index_free (bool) -- optional, the Galactic diffuse is given a power-law spectral shape but the by default the index is frozen, setting this flag to True allows that to be free for additional freedom in diffuse fit
+	#use_old_names (bool) -- optional, flag to use the naming convention from make1FGLxml.py and make2FGLxml.py with a leading underscore and no spaces
+	def make_model(self,galactic_file="$(FERMI_DIR)/refdata/fermi/galdiffuse/gll_iem_v07.fits",galactic_name='gll_iem_v07',
+                      isotropic_file="$(FERMI_DIR)/refdata/fermi/galdiffuse/iso_P8R3_SOURCE_V3_v1.txt",isotropic_name='iso_P8R3_SOURCE_V3_v1',
+                      norms_free_only=False,extended_directory=None,free_radius=-1,max_free_radius=None,extra_radius=10,sigma_to_free=5,
+                      variable_free=True,force_point_sources=False,extended_catalog_names=False,make_region=True,region_file=None,
+                      galactic_index_free=False,use_old_names=False):
 
-def addSrcsXML(sL,GD,GDn,ISO,ISOn,oldNames=False):
-	varValue=sL.VarLim
+                self.galactic_file=galactic_file
+                self.galactic_name=galactic_name
+
+                self.isotropic_file=isotropic_file
+                self.isotropic_name=isotropic_name
+                
+		self.free_radius=(self.roi[2] if free_radius<=0 else free_radius)
+		self.max_free_radius=(self.free_radius if max_free_radius is None else max_free_radius)
+		if self.max_free_radius<self.free_radius:
+                        warnings.warn(f'max_free_radius ({max_free_radius:.1f}) is < free_radius ({free_radius:.1f}), making the parameter useless.')
+                self.extra_radius=extra_radius
+
+		self.variable_free=variable_free
+		self.force_point_sources=force_point_sources
+		self.extended_catalog_names=extended_catalog_names
+		self.norms_free_only=norms_free_only
+
+		self.extended_directory=extended_directory if extended_directory is None else\
+                                    os.sep.join(['$(FERMI_DIR)','data','pyBurstAnalysisGUI','templates'])
+		
+		self.sigma_to_free=sigma_to_free
+		self.make_region=make_region
+		self.galactic_index_free=galactic_index_free
+
+                self.create_region_model()
+		
+		if make_region:
+                        if region_file is None:
+                                region_file=os.path.basename(self.output_name).split('.')[:-1]
+                                region_file=reduce(lambda s1,s2:s1+s2,region_file).join(['ROI_','.reg'])
+                                
+                         self.region_file=os.path.join(self.write_directory,region_file)
+                        
+                        print('Building ds9-style region file',end='...')
+                        build_region(self.region_file,self.sources)
+                        print(f'done!\nFile saved as {self.region_file}.')
+        
+        def create_region_model(self):
+		print(f'Creating spatial and spectral model from the 4FGL DR-{self.DR} catalog: {self.catalog_file}.')
+		
+		if os.path.basename(self.catalog_file).split('.')[-1].lower()=='xml':
+                        self.get_sources_xml()
+		else:
+                        self.get_sources_fits()
+
+                #from either method we get a self.sources attribute which is a nested dictionary with all the sources
+                self.build_model()
+
+        def build_model(self):
+                #keep track of point and extended sources, similar to before
+                self.num_extended=0
+                self.num_point=0
+                #use the self.sources nested dictionary to create the output xml file
+                output_xml=minidom.getDOMImplementation().createDocument(None,'source_library',None)
+                output_xml.documentElement.setAttribute('title','source library')
+                for source_name in self.sources.keys():
+                        source_out=output_xml.createElement('source')
+                        source_out.setAttribute('name','_'+source_name if self.use_old_names else source_name)
+                        source_out.setAttribute('ROI_Center_Distance',"{self.sources[source_name]['roi_distance']:.2f}")
+                        source_out.setAttribute('type','DiffuseSource' if self.sources[source_name]['Extended'] else 'PointSource')
+                        if self.sources[source_name]['Extended']:
+                                self.num_extended+=1
+                        else:
+                                self.num_point+=1
+                        
+                        #need to determine if parameters should be free or fixed
+                        #first, basic checks for conditions meaning everything is fixed
+                        if self.sources[source_name]['roi_distance']>self.max_free_radius or\
+                           (self.sources[source_name]['roi_distance']>self.free_radius and not self.sources[source_name]['variable']) or\
+                           (self.sources[source_name]['roi_distance']<=self.free_radius and not self.sources[source_name]['significant']\
+                            and not self.sources[source_name]['variable']):
+
+                                for spectrum_par in self.sources[source_name]['spectrum'].keys():
+                                        if spectrum_par not in ['model','spectrum_file']:
+                                                self.sources[source_name]['spectrum'].update([(spectrum_par+'_free',False)])
+
+                                self.sources[source_name].update([('free',False)])
+
+                        #now, check if the source is far from the ROI center but variable
+                        #or if close by but not formally significant or close by and significant but norms_only_free set
+                        #free the normalization, only
+                        elif (self.sources[source_name]['roi_distance']<=self.max_free_radius and\
+                             self.sources[source_name]['roi_distance']>self.free_radius and self.sources[source_name]['variable']) or \
+                             (self.sources[source_name]['roi_distance']<=self.free_radius and not self.sources[source_name]['significant']\
+                              and self.sources[source_name]['variable']) or (self.sources[source_name]['roi_distance']<=self.free_radius\
+                                and self.sources[source_name]['significant'] and self.norms_free_only):
+                                
+                                for spectrum_par in self.sources[source_name]['spectrum'].keys():
+                                        if spectrum_par  not in ['model','spectrum_file']:
+                                                if spectrum_par in ['Prefactor','Integral','norm','Normalization']:
+                                                        self.sources[source_name]['spectrum'].update([(spectrum_par+'_free',True)])
+                                                else:
+                                                        self.sources[source_name]['spectrum'].update([(spectrum_par+'_free',False)])
+
+                                self.sources[source_name].update([('free':True)])
+
+                        #now, check if the source is close to the ROI center and significant
+                        #need to also consider if norms_free_only flag is set
+                        #this should be the only possiblity left, but for now we'll use elif and put an else
+                        #statement to catch things we haven't thought of, remove after successful testing
+                        elif self.sources[source_name]['roi_distance']<=self.free_radius and self.sources[source_name]['significant']:
+                                for spectrum_par in source.sources[source_name]['spectrum'].keys():
+                                        if spectrum_par not in ['model','spectrum_file']:
+                                                if spectrum_par in ['Scale','Index2','Eb']:
+                                                        self.sources[source_name]['spectrum'].update([(spectrum_par+'_free',False)])
+                                                else:
+                                                        self.sources[source_name]['spectrum'].update([(spectrum_par+'_free',True)])
+
+                                self.sources[source_name].update([('free':True)])
+                        #the following else statement is for testing only
+                        #should not drop into that section of code, remove after many successful tests
+                        #replace the 'elif' above with simply 'else'
+                        else:
+                                print('Should not have arrived here! Look at nested dictionary to see what happened.')
+                                print(self.sources[source_name])
+                                raise RuntimeError()
+
+                        #now we need to add the extended_directory to extended source file names
+                        if self.sources[source_name]['Extended'] and\
+                           self.sources[source_name]['spatial']['spatial_model'] in ['SpatialMap','MapCubeFunction']:
+                                self.sources[source_name]['spatial']['spatial_file']=os.path.join(self.extended_directory,
+                                                                                self.sources[source_name]['spatial']['spatial_file'])
+                                print(f'Extended source {source_name} in model, make sure that the following path to extended template is correct:',end=' ')
+                                print(self.sources[source_name]['spatial']['spatial_file'])
+
+                        #now that we have added flags for spectral parameters to be free or fixed
+                        #and possibly adjusted extended source file paths
+                        #let's create the spectrum and spatialModel elements
+                        #and then add them to the source and add that to the larger XML document
+                        spectrum=Spectrum(**self.sources[source_name]['spectrum'])
+                        source_out.appendChild(spectrum.spectrum)
+
+                        spatial=Spatial(**self.sources[source_name]['spatial'])
+                        source_out.appendChild(spatial.spatial)
+
+                        #now add the source to the output_xml
+                        output_xml.documentElement.appendChild(source_out)
+
+                #we've cycled through all of the sources, now we need to add the diffuse components
+                galactic_spectrum=Spectrum(model='PowerLaw',Prefactor=1,Index=0,Scale=100,Prefactor_free=True,
+                                           Index_free=self.galactic_index_free)
+                galactic_spatial=Spatial(spatial_model='MapCubeFunction',spatial_file=self.galactic_model_file)
+
+                isotropic_spectrum=SPectrum(model='FileFunction',spectrum_file=self.isotropic_file,apply_edisp='false')
+                isotropic_spatial=Spatial(spatial_model='ConstantValue')
+
+                Galactic=output_xml.createElement('source')
+                Galactic.setAttribute('name',self.galactic_name)
+                Galactic.setAttribute('type','DiffuseSource')
+
+                Galactic.appendChild(galactic_spectrum.spectrum)
+                Galactic.appendChild(galactic_spatial.spatial)
+
+                Isotropic=output_xml.createElement('source')
+                Isotropic.setAttribute('name',self.isotropic_name)
+                Isotropic.setAttribute('type','DiffuseSource')
+
+                Isotropic.appendChild(isotropic_spectrum.spectrum)
+                Isotropic.appendChild(isotropic_spatial.spatial)
+
+                output_xml.documentElement.appendChild(Galactic)
+                output_xml.documentElement.appendChild(Isotropic)
+
+                out_string=filter(lambda s: len(s) and not s.isspace(),
+                                  output_xml.toprettyxml(' ').splitlines(True))
+
+                with open(self.output_name,'w') as output_file:
+                        outfile.write(''.join(out_string))
+                if self.force_point_sources:
+                        print(f'Added {self.num_point} point sources, note that any extended sources',end=' ')
+                        print(f'(except diffuse components) have been modeled as point sources.')
+                else:
+                        print(f'Added {self.num_point} point sources and {self.num_extended} sources.')
+                        if self.num_extended>0:
+                                print('If you plan to analyze LAT data with unbinned likelihood,',end=' ')
+                                print('you will need to run gtdiffrsp for the extended sources',end=' ')
+                                print('or rerun makeModel with force_point_sources set to True.')
+
+
+        def get_sources_xml(self):
+                catalog=minidom.parse(self.catalog_file)
+                sources=catalog.getElementsByTagName('source')
+
+                #traverse through the sources once, just to get positional information
+                right_ascensions,declinations=[],[]
+                for source in sources:
+                        try:
+                                right_ascensions.append(float(source.getAttribute('RA')))
+                                declinations.append(float(source.getAttribute('DEC')))
+                        except:
+                                for parameter in source.getElementsByTagName('spatialModel')[0].getElementsByTagName('parameter'):
+                                        if parameter.getAttribute('name')=='RA':
+                                                right_ascensions.append(float(parameter.getAttribute('value')))
+                                        elif parameter.getAttribute('name')=='DEC':
+                                                declinations.append(float(parameter.getAttribute('value')))
+
+                right_ascensions=np.array(right_ascensions)
+                declinations=np.array(declinations)
+                
+                self.source_distances=angular_separation(self.ROI[0],self.ROI[1],
+                                                         right_ascensions,declinations)
+
+                source_mask=self.source_distances<=self.ROI[2]+self.extra_radius
+                self.source_distances=self.source_distances[source_mask]
+                distance_index=self.source_distances.argsort()
+                self.source_distances=self.source_distances[distance_index]
+                
+                sources=sources[source_mask][distance_index]
+                right_ascensions=right_ascentions[source_mask][distance_index]
+                declinations=declinations[source_mask][distance_index]
+
+                #now traverse the shorter list of sources to create the self.sources nested dictionary
+                self.sources={}
+                for source,RA,DEC,distance in zip(self.sources,right_ascensions,declinations,self.source_distances):
+                        name=source.getAttribute('name')
+                        self.sources.update([(name,{'roi_distance':distance,'RA':RA,'DEC':DEC})])
+
+                        if source.getAttribute('type')=='DiffuseSource':
+                                self.sources[name].update([('Extended',not self.force_point_sources)])
+
+                        #evaluate variability and significance threshold info
+                        self.sources[name].update([('variable',float(src.getAttribute('Variability_Index'))>=self.variability_threshold),
+                                ('significant',float(source.getAttribute('TS_value'))>=self.sigma_to_free)])
+
+                        #now get the spectrum information
+                        spectrum=source.getElementsByTagName('spectrum')
+                        
+                        self.sources[name].update([('spectrum',
+                             {'model':specrum[0].getAttribute('type')})])
+                        
+                        for parameter in spectrum[0].getElementsByTagName('parameter'):
+                        	self.sources[name]['spectrum'].update([(parameter.getAttribute('name'),
+                        	float(parameter.getAttribute('value'))*float(parameter.getAttribute('scale')))])
+                        
+                        #now for the spatial information
+                        spatial=source.getElementsByTagName('spatialModel')
+                        
+                        if self.sources[name]['Extended']:
+                        	self.sources[name].update([('spatial',
+                                        {'spatial_model':spatial[0].getAttribute('type')})])
+                        	for parameter in spatial[0].getElementsByTagName('parameter'):
+                        		self.sources[name]['spatial'].update([(parameter.getAttribute('name'),
+                        		parameter.getAttribute('value'))])
+                        
+                        else:
+                        	self.sources[name].update([('spatial',
+                        	    {'spatial_model':'SkyDir','RA':RA,'DEC':DEC})])
+                        
+                        
+
+        def get_sources_fits(self):
+                #first, open the catalog FITS file and extract the information
+                #we care about, cast as numpy arrays for ease later
+                with pyfits.open(self.catalog_file) as catalog:
+                        source_info=catalog['LAT_Point_Source_Catalog'].data.field
+                        extended_info=catalog['ExtendedSources'].data.field
+                        
+                        extended_names=np.array(extended_info('Source_Name'))
+                        extended_files=np.array(extended_info('Spatial_Filename'))
+                	extended_spatial_functions=np.array(extended_info('Spatial_Function'))
+                	extended_extents=np.array(extended_info('Model_SemiMajor'))
+                        extended_RA=np.array(extended_info('RAJ2000'))
+                	extended_DEC=np.array(extended_info('DEJ2000'))
+
+                        #cast these as a data frame with the extended_name as the index
+                	#making it much easier to search later using just .loc
+                	#instead of a nested for loop
+                	extended_sources=pd.DataFrame(np.c_[extended_files,extended_spatial_functions,
+                                extended_extents,extended_RA,extended_DEC],
+                                columns=['file','function','extent','RA','DEC'],
+                                index=extended_names)
+                	
+                	names=np.array(source_info('Source_Name'))
+                	
+                	average_significances=np.array(source_info('Signif_Avg'))
+                	variability_indices=np.array(source_info('Variability_Index'))
+                	
+                	extended_source_names=np.array(source_info('Extended_Source_Name'))
+                	
+                	right_ascensions=np.array(source_info('RAJ2000'))
+                	declinations=np.array(source_info('DEJ2000'))
+                	
+                	power_law_fluxes=np.array(source_info('PL_Flux_Density'))
+                	log_parabola_fluxes=np.array(source_info('LP_Flux_Density'))
+                	cutoff_fluxes=np.array(source_info('PLEC_Flux_Density'))
+                	
+                	pivot_energies=np.array(source_info('Pivot_Energy'))
+                	
+                	power_law_indices=np.array(source_info('PL_Index'))
+                	
+                	log_parabola_indices=np.array(source_info('LP_Index'))
+                	log_parabola_betas=np.array(source_info('LP_beta'))
+                	
+                        cutoff_indices=np.array(source_info(('PLEC_IndexS' if self.DR==3 else 'PLEC_Index')))
+                        cutoff_expfactors=np.array(source_info(('PLEC_ExpfactorS' if self.DR==3 else 'PLEC_Expfactor')))
+                        cutoff_expindices=np.array(source_info('PLEC_Exp_Index'))
+                        
+                        spectral_types=np.array(source_info('SpectrumType'))
+
+                #calculate the source distances with respect to the ROI center
+                self.source_distances=angular_separation(self.ROI[0],self.ROI[1],
+                                                         right_ascensions,declinations)
+
+                #get a mask for sources in the distance range we care about
+                #and then sort them in ascending order
+                source_mask=self.source_distances<=self.ROI[2]+self.extra_radius
+                self.source_distances=self.source_distances[source_mask]
+                distance_index=self.source_distances.argsort()
+                self.source_distances=self.source_distances[distance_index]
+
+                #now, downselect all the other arrays
+                right_ascensions=right_ascentions[source_mask][distance_index]
+                declinations=declinations[source_mask][distance_index]
+                
+                names=names[source_mask][distance_index]
+                
+                average_significances=average_significances[source_mask][distance_index]
+                variability_indices=variability_indices[source_mask][distance_index]
+                
+                extended_source_names=extended_source_names[source_mask][distance_index]
+                
+                power_law_fluxes=power_law_fluxes[source_mask][distance_index]
+                log_parabola_fluxes=log_parabola_fluxes[source_mask][distance_index]
+                cutoff_fluxes=cutoff_fluxes[source_mask][distance_index]
+                
+                pivot_energies=pivot_energies[source_mask][distance_index]
+                
+                power_law_indices=power_law_indices[source_mask][distance_index]
+                
+                log_parabola_indices=log_parabola_indices[source_mask][distance_index]
+                log_parabola_betas=log_parabola_betas[source_mask][distance_index]
+                
+                cutoff_indices=cutoff_indices[source_mask][distance_index]
+                cutoff_expfactors=cutoff_expfactors[source_mask][distance_index]
+                cutoff_expindices=cutoff_expindices[source_mask][distance_index]
+                
+                spectral_types=spectral_types[source_mask][distance_index]
+
+                self.sources={}
+
+                for idx in range(len(names)):
+                        #check if the source is extended or not, decide on name for source
+                        #and if it is extended
+                        if names[idx][-1]=='e':
+                                name=(names[idx] if self.extended_catalog_names else extended_source_names[idx])
+                                self.sources.update([(name,{'roi_distance':self.source_distances[idx]})])
+                                self.sources[name].update([('Extended',not self.force_point_sources)])
+                        else:
+                                name=names[idx]
+                                self.sources.update([(name,{'roi_distance':self.source_distances[idx]})])
+                                self.sources[name].update([('Extended',False)])
+                        
+                        #evaluate variability and significance threshold info
+                        #as well as the RA and DEC to be used if building a .reg file
+                        self.sources[name].update([('variable',variability_indices[idx]>=self.variability_threshold),
+                                                 ('significant',average_significances[idx]>=self.sigma_to_free),
+                                                 ('RA',right_ascensions[idx]),('DEC',declinations[idx])])
+
+                        #now get source spectral information in the form we need
+                        #I need to verify that PowerLaw, LogParabola, and PLSuperExpCutoff4
+                        #are the only models we have in 4FGL DR 1, 2, and 3
+                        if spectral_types[idx]=='PowerLaw':
+                                self.sources[name].update([('spectrum',
+                                        {'model':'PowerLaw',
+                                         'Prefactor':power_law_fluxes[idx],
+                                         'Index':power_law_indices[idx],
+                                         'Scale':pivot_energies[idx]})])
+
+                        elif spectral_types[idx]=='LogParabola':
+                                self.sources[name].update([('spectrum',
+                                        {'model':'LogParabola',
+                                         'norm':log_parabola_fluxes[idx],
+                                         'alpha':log_prabola_indices[idx],
+                                         'beta':log_parabola_betas[idx],
+                                         'Eb':pivot_energies[idx]})])
+
+                        else:
+                                if self.DR==3:
+                                        self.sources[name].update([('spectrum',
+                                                {'model':'PLSuperExpCutoff4',
+                                                 'Prefactor':cutoff_fluxes[idx],
+                                                 'IndexS':cutoff_indices[idx],
+                                                 'Scale':pivot_energies[idx],
+                                                 'ExpfactorS':cutoff_expfactors[idx],
+                                                 'Index2':cutoff_expindices[idx]})])
+
+                        #now for the spatial information
+                        #check if extended, this flag already uses force_point_sources flag
+                        if self.sources[name]['Extended']:
+                                #do extended stuff
+                                row=extended_sources.loc[extended_source_names[idx]]
+                                #if one of the 'Radial' models
+                                if row.function[:6]=='Radial':
+                                        self.sources[name].update([('spatial',
+                                                {'spatial_model':'RadialGaussian' if row.function=='RadialGauss' else row.function,
+                                                 'RA':row.RA,
+                                                 'DEC':row.DEC})])
+                                        if row.function=='RadialDisk':
+                                                self.sources[name]['spatial'].update([('Radius',row.extent)])
+                                        else:
+                                                self.sources[name]['spatial'].update([('Sigma',row.extent/(-2.np.log(0.32))**0.5)])
+
+                                #otherwise it is a spatial map
+                                else:
+                                        self.sources[name].update([('spatial',
+                                                {'spatial_model':'SpatialMap',
+                                                 'spatial_file':row.file})])
+
+                        #point source
+                        else:
+                                self.sources[name].update([('spatial',
+                                        {'type':'SkyDir',
+                                         'RA':right_ascensions[idx],
+                                         'DEC':declinations[idx]})])
+
+                #done with the loop, object will use the sources dictionary to build the model
+
+
+
+                        
+
+def addSrcsXML(sL,GD,GDn,ISO,ISOn,use_old_names=False):
+	varValue=sL.variability_threshold
 	inputXml=minidom.parse(sL.srcs)
 	outputXml=minidom.getDOMImplementation().createDocument(None,'source_library',None)
 	outputXml.documentElement.setAttribute('title','source library')
@@ -128,7 +532,7 @@ def addSrcsXML(sL,GD,GDn,ISO,ISOn,oldNames=False):
 			Ext=(True if (src.getAttribute('type')=='DiffuseSource' and not sL.psF) else False)
 			sname=src.getAttribute('name')
 			
-			if oldNames:#if you want the same naming convention as in make1FGLxml.py and make2FGLxml.py, e.g., preceeded by an underscore and no spaces
+			if use_old_names:#if you want the same naming convention as in make1FGLxml.py and make2FGLxml.py, e.g., preceeded by an underscore and no spaces
 				sn='_'
 				for N in str(sname).split(' '):
 					sn+=N
@@ -144,11 +548,11 @@ def addSrcsXML(sL,GD,GDn,ISO,ISOn,oldNames=False):
 			srcOut=outputXml.createElement('source')
 			srcOut.setAttribute('name',sname)
 			srcOut.setAttribute('ROI_Center_Distance',"%.2f"%dist)
-			if dist>=sL.roi[2] or dist>=sL.maxRad:
+			if dist>=sL.roi[2] or dist>=sL.max_free_radius:
 				Sources[sname]['free']=False
 				for p in specPars:
 				  specOut.appendChild(parameter_element("0","%s"%str(p.getAttribute('name')),"%s"%str(p.getAttribute('max')),"%s"%str(p.getAttribute('min')),"%s"%str(p.getAttribute('scale')),"%s"%str(p.getAttribute('value'))))
-			elif dist>sL.radLim:
+			elif dist>sL.free_radius:
 				if sL.var and varIdx>=varValue:
 					Sources[sname]['free']=True
 					for p in specPars:
@@ -244,15 +648,15 @@ def addSrcsXML(sL,GD,GDn,ISO,ISOn,oldNames=False):
 	if not sL.psF:
 		print('Added %i point sources and %i extended sources'%(ptSrcNum,extSrcNum))
 		if extSrcNum>0:
-			print('If using unbinned likelihood you will need to rerun gtdiffrsp for the extended sources or rerun the makeModel function with optional argument psForce=True')
+			print('If using unbinned likelihood you will need to rerun gtdiffrsp for the extended sources or rerun the makeModel function with optional argument force_point_sources=True')
 	else:
-		print('Added %i point sources, note that any extended sources in ROI were modeled as point sources becaue psForce option was set to True'%ptSrcNum)
+		print('Added %i point sources, note that any extended sources in ROI were modeled as point sources becaue force_point_sources option was set to True'%ptSrcNum)
 	if sL.reg:
 		BuildRegion(sL,Sources)
 	return
 
 #function to cycle through the source list and add point source entries
-def addSrcsFITS(sL,GD,GDn,ISO,ISOn,oldNames):
+def addSrcsFITS(sL,GD,GDn,ISO,ISOn,use_old_names):
 	model=open(sL.out,'w') #open file in write mode, overwrites other files of same name
 	catfile=pyfits.open(sL.srcs) #open source list file and access necessary fields, requires LAT source catalog definitions and names
 	data=catfile['LAT_Point_Source_Catalog'].data
@@ -288,7 +692,7 @@ def addSrcsFITS(sL,GD,GDn,ISO,ISOn,oldNames):
 	model.write('<?xml version="1.0" ?>\n')
 	model.write('<source_library title="source library">\n')
 	model.write('\n<!-- Point Sources -->\n')
-	step=(sL.roi[2]+sL.ER)/5. #divide ROI radius plus ExtraRadius degrees into 5 steps for ordering of sources
+	step=(sL.roi[2]+sL.ER)/5. #divide ROI radius plus extra_radiusius degrees into 5 steps for ordering of sources
 	i=1
 	radii=[]
 	ptSrcNum=0
@@ -316,12 +720,12 @@ def addSrcsFITS(sL,GD,GDn,ISO,ISOn,oldNames):
 					extSrcNum+=1
 					Name='<source ROI_Center_Distance="%.3f" name="%s" type="DiffuseSource">\n' %(dist,En)
 				else:
-					if E and not sL.E2C:#even if forcing all to point sources, use extended name except if E2CAT flag is set
+					if E and not sL.E2C:#even if forcing all to point sources, use extended name except if extended_catalog_names flag is set
 						Sources[En]={'ra':r,'dec':d,'stype':t,'E':E}
 						Name='<source ROI_Center_Distance="%.3f" name="%s" type="PointSource">\n' %(dist,En)
 					else:
 						Sources[n]={'ra':r,'dec':d,'stype':t,'E':E}
-						if oldNames:
+						if use_old_names:
 							srcname='_'
 							for N in n.split(' '):
 								srcname+=N
@@ -393,9 +797,9 @@ def addSrcsFITS(sL,GD,GDn,ISO,ISOn,oldNames):
 	if not sL.psF:
 		print('Added %i point sources and %i extended sources'%(ptSrcNum,extSrcNum))
 		if extSrcNum>0:
-			print('If using unbinned likelihood you will need to rerun gtdiffrsp for the extended sources or rerun the makeModel function with optional argument psForce=True')
+			print('If using unbinned likelihood you will need to rerun gtdiffrsp for the extended sources or rerun the makeModel function with optional argument force_point_sources=True')
 	else:
-		print('Added %i point sources, note that any extended sources in ROI were modeled as point sources becaue psForce option was set to True'%ptSrcNum)
+		print('Added %i point sources, note that any extended sources in ROI were modeled as point sources becaue force_point_sources option was set to True'%ptSrcNum)
 	#add galactic diffuse with PL spectrum, fix index to zero for general use, those who want it to be free can unfreeze parameter manually
 	model.write('\n<!-- Diffuse Sources -->\n')
 	Name='\n<source name="%s" type="DiffuseSource">\n' %GDn
@@ -432,286 +836,6 @@ def addSrcsFITS(sL,GD,GDn,ISO,ISOn,oldNames):
 		BuildRegion(sL,Sources)
 	return
 
-def BuildRegion(sL,Sources):
-	myreg=open(sL.regFile,'w')#note that this will overwrite previous region files of the same name
-	myreg.write('# Region File format: DS9 version ?')#I don't actually know, but I think it's one of the later ones, need to verify
-	myreg.write('\n# Created by make3FGLxml.py')
-	myreg.write('\nglobal font="roman 10 normal" move =0')
-	for k in Sources.keys():
-		src=Sources[k]
-		#get color based on if the source is free or not
-		color=('green' if src['free'] else 'magenta')
-		if src['E']:#if the source is extended, always have the point be a "big" box
-			myreg.write('\nJ2000;point(%.3f,%.3f) # point = box 18 color = %s text={%s}'%(src['ra'],src['dec'],color,k))
-		else:#if the source is a point source, choose the point type based on spectral model
-			ptype=('cross' if (src['stype']=='PLSuperExpCutoff' or src['stype']=='PLSuperExpCutoff2') else 'diamond' if src['stype']=='LogParabola' else 'circle')
-			myreg.write('\nJ2000;point(%.3f,%.3f) # point = %s 15 color = %s text={%s}'%(src['ra'],src['dec'],ptype,color,k))
-	myreg.close()
-	return
-
-def PLspec(sL,f,i,p,dist,TS,vi):
-	varValue=sL.VarLim
-	fscale=int(floor(log10(f)))
-	spec='\t<spectrum type="PowerLaw">\n'
-	spec+='\t<!-- Source is %s degrees away from ROI center -->\n' %dist
-	if dist>sL.roi[2]: #if beyond ROI, shouldn't attempt to fit parameters
-		spec+='\t<!-- Source is outside ROI, all parameters should remain fixed -->\n'
-		spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index" scale="-1.0" value="%s"/>\n' %i
-		free=False
-	elif(dist>sL.radLim):
-		if dist>sL.maxRad:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		elif vi<varValue or not sL.var:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source is outside specified radius limit of %s but variability index %.2f is greater than %.2f and varFree set to True-->\n'%(sL.radLim,vi,varValue)
-			free=True
-			spec+='\t\t<parameter free="1" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index" scale="-1.0" value="%s"/>\n' %i
-	elif(TS<sL.sig):
-		if vi<varValue or not sL.var:
-			spec+='\t<!-- Source signficance %.1f is less than specified minimum for a free source of %s -->\n'%(TS,sL.sig)
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source significane %.1f is less than specified minimum for a free source of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(TS,sL.sig,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index" scale="-1.0" value="%s"/>\n' %i
-	else:
-		spec+='\t\t<parameter free="1" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		free=True
-		if sL.nO:
-			spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index" scale="-1.0" value="%s"/>\n' %i
-		else:
-			spec+='\t\t<parameter free="1" max="10.0" min="0.0" name="Index" scale="-1.0" value="%s"/>\n' %i
-	spec+='\t\t<parameter free="0" max="5e5" min="30" name="Scale" scale="1.0" value="%f"/>\n' %p
-	spec+='\t</spectrum>\n'
-	return spec,free
-
-def CO2spec(sL,f,i,p,a,ei,dist,TS,vi):
-	varValue=sL.VarLim
-	f*=exp(a*(p**ei))
-	fscale=int(floor(log10(f)))
-	spec='\t<spectrum type="PLSuperExpCutoff2">\n'
-	spec+='\t<!-- Source is %s degrees away from ROI center -->\n' %dist
-	#i=(i if i>=0 else 2.)#some pulsars with index1 < 0 assuming standard convention, means rising counts spectrum at low E, odd
-	if(dist>sL.roi[2]): #if beyond ROI, shouldn't attempt to fit parameters
-		spec+='\t<!-- Source is outside ROI, all parameters should remain fixed -->\n'
-		free=False
-		spec+='\t\t<parameter free="0" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index1" scale="-1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="100" min="-1" name="Expfactor" scale="0.001" value="%f"/>\n'%(a*1000.)
-	elif(dist>sL.radLim):
-		if dist>sL.maxRad:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		elif vi<varValue or not sL.var:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source is outside specified radius limit of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(sL.radLim,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index1" scale="-1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="100" min="-1" name="Expfactor" scale="0.001" value="%f"/>\n'%(a*1000.)
-	elif(TS<sL.sig):
-		if vi<varValue or not sL.var:
-			spec+='\t<!-- Source significance %.1f is less than specified minimum for a free source of %s -->\n'%(TS,sL.sig)
-			spec+='\t\t<parameter free="0" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source significance %.1f is less than specified minimum for a free source of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(TS,sL.sig,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index1" scale="-1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="100" min="-1" name="Expfactor" scale="0.001" value="%f"/>\n'%(a*1000.)
-	else:
-		free=True
-		spec+='\t\t<parameter free="1" max="1e4" min="1e-4" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		if sL.nO:
-			spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="Index1" scale="-1.0" value="%s"/>\n' %i
-			spec+='\t\t<parameter free="0" max="100" min="-1" name="Expfactor" scale="0.001" value="%f"/>\n'%(a*1000.)
-		else:
-			spec+='\t\t<parameter free="1" max="10.0" min="0.0" name="Index1" scale="-1.0" value="%s"/>\n' %i
-			spec+='\t\t<parameter free="1" max="100" min="-1" name="Expfactor" scale="0.001" value="%f"/>\n'%(a*1000.)
-	spec+='\t\t<parameter free="0" max="5e5" min="30" name="Scale" scale="1.0" value="%f"/>\n'%p
-	spec+='\t\t<parameter free="0" max="5" min="0" name="Index2" scale="1.0" value="%f"/>\n'%ei
-	spec+='\t</spectrum>\n'
-	return spec,free
-
-def CO4spec(sL,f,i,p,a,ei,dist,TS,vi):
-	varValue=sL.VarLim
-	fscale=int(floor(log10(f)))
-	spec='\t<spectrum type="PLSuperExpCutoff4">\n'
-	spec+='\t<!-- Source is %s degrees away from ROI center -->\n' %dist
-	if(dist>sL.roi[2]): #if beyond ROI, shouldn't attempt to fit parameters
-		spec+='\t<!-- Source is outside ROI, all parameters should remain fixed -->\n'
-		free=False
-		spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="IndexS" scale="-1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="100" min="-10" name="ExpfactorS" scale="0.1" value="%f"/>\n'%(a*10.)
-	elif(dist>sL.radLim):
-		if dist>sL.maxRad:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		elif vi<varValue or not sL.var:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source is outside specified radius limit of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(sL.radLim,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="IndexS" scale="-1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="100" min="-10" name="ExpfactorS" scale="0.1" value="%f"/>\n'%(a*10.)
-	elif(TS<sL.sig):
-		if vi<varValue or not sL.var:
-			spec+='\t<!-- Source significance %.1f is less than specified minimum for a free source of %s -->\n'%(TS,sL.sig)
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source significance %.1f is less than specified minimum for a free source of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(TS,sL.sig,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="IndexS" scale="-1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="100" min="-10" name="ExpfactorS" scale="0.1" value="%f"/>\n'%(a*10.)
-	else:
-		free=True
-		spec+='\t\t<parameter free="1" max="1e6" min="0" name="Prefactor" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		if sL.nO:
-			spec+='\t\t<parameter free="0" max="10.0" min="0.0" name="IndexS" scale="-1.0" value="%s"/>\n' %i
-			spec+='\t\t<parameter free="0" max="100" min="-10" name="ExpfactorS" scale="0.1" value="%f"/>\n'%(a*10.)
-		else:
-			spec+='\t\t<parameter free="1" max="10.0" min="0.0" name="IndexS" scale="-1.0" value="%s"/>\n' %i
-			spec+='\t\t<parameter free="1" max="100" min="-10" name="ExpfactorS" scale="0.1" value="%f"/>\n'%(a*10.)
-	spec+='\t\t<parameter free="0" max="5e5" min="30" name="Scale" scale="1.0" value="%f"/>\n'%p
-	spec+='\t\t<parameter free="0" max="5" min="0" name="Index2" scale="1.0" value="%f"/>\n'%ei
-	spec+='\t</spectrum>\n'
-	return spec,free
-
-def LPspec(sL,f,i,p,b,dist,TS,vi,fixAll=False):
-	varValue=sL.VarLim
-	fscale=int(floor(log10(f)))
-	spec='\t<spectrum type="LogParabola">\n'
-	spec+='\t<!-- Source is %s degrees away from ROI center -->\n' %dist
-	if(dist>sL.roi[2]): #if beyond ROI, shouldn't attempt to fit parameters
-		spec+='\t<!-- Source is outside ROI, all parameters should remain fixed -->\n'
-		spec+='\t\t<parameter free="0" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		spec+='\t\t<parameter free="0" max="5.0" min="0.0" name="alpha" scale="1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="10.0" min="-5.0" name="beta" scale="1.0" value="%s"/>\n'%b
-		free=False
-	elif fixAll: #if set to fix all, should be only for crab IC component
-		spec+='\t<!-- Source is kept fixed in 4FGL-DR3 analysis -->\n'
-		spec+='\t\t<parameter free="0" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		spec+='\t\t<parameter free="0" max="5.0" min="0.0" name="alpha" scale="1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="10.0" min="-5.0" name="beta" scale="1.0" value="%s"/>\n'%b
-		free=False
-	elif(dist>sL.radLim):
-		if dist>sL.maxRad:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		elif vi<varValue or not sL.var:
-			spec+='\t<!-- Source is outside specified radius limit of %s -->\n'%sL.radLim
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source is outside specified radius limit of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(sL.radLim,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="5.0" min="0.0" name="alpha" scale="1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="10.0" min="-5.0" name="beta" scale="1.0" value="%s"/>\n'%b
-	elif(TS<sL.sig):
-		if vi<varValue or not sL.var:
-			spec+='\t<!-- Source significance %.1f is less than specified minimum for a free source of %s -->\n'%(TS,sL.sig)
-			spec+='\t\t<parameter free="0" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=False
-		else:
-			spec+='\t<!-- Source significance %.1f is less than specified minimum for a free source of %s but variability index %.2f is greater than %.2f and varFree is set to True -->\n'%(TS,sL.sig,vi,varValue)
-			spec+='\t\t<parameter free="1" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-			free=True
-		spec+='\t\t<parameter free="0" max="5.0" min="0.0" name="alpha" scale="1.0" value="%s"/>\n' %i
-		spec+='\t\t<parameter free="0" max="10.0" min="-5.0" name="beta" scale="1.0" value="%s"/>\n'%b
-	else:
-		free=True
-		spec+='\t\t<parameter free="1" max="1e6" min="0" name="norm" scale="1e%i" value="%s"/>\n' %(fscale,f/10**fscale)
-		if sL.nO:
-			spec+='\t\t<parameter free="0" max="5.0" min="0.0" name="alpha" scale="1.0" value="%s"/>\n' %i
-			spec+='\t\t<parameter free="0" max="10.0" min="-5.0" name="beta" scale="1.0" value="%s"/>\n'%b
-		else:
-			spec+='\t\t<parameter free="1" max="5.0" min="0.0" name="alpha" scale="1.0" value="%s"/>\n' %i
-			spec+='\t\t<parameter free="1" max="10.0" min="-5.0" name="beta" scale="1.0" value="%s"/>\n'%b
-	spec+='\t\t<parameter free="0" max="5e5" min="30" name="Eb" scale="1.0" value="%s"/>\n'%p
-	spec+='\t</spectrum>\n'
-	return spec,free
-
-#this function searches the header of the ft1 to find the Position keyword and extract the ra and dec values
-def getPos(ft1):
-	file=pyfits.open(ft1)
-	num=file[1].header['NDSKEYS']
-	header=file[1].header
-	right='POS(RA,DEC)'
-	i=1
-	keynum=0
-	while i<=num:  #this step is necessary since it is not clear that the POS key word will have the same number always
-		word='DSTYP%i' %i
-		test=file[1].header[word]
-		if(test==right):
-			keynum=i
-			i=num
-		i+=1
-	if(keynum==0):  #DSKEYS start numbering at 1, if this value hasn't been updated, KEYword doesn't exist
-		print('Error: No position keyword found in fits header (assuming position is RA and DEC.  Exiting...')
-		exit()
-	keyword='DSVAL%i' %keynum
-	try:
-		ra,dec,rad=header[keyword].strip('CIRCLE()').split(',') #gets rid of the circle and parenthesis part and splits around the comma
-		float(ra)
-	except:
-		ra,dec,rad=header[keyword].strip('circle()').split(',')
-	file.close()
-	return float(ra),float(dec),float(rad)
-	
-#calculates the angular separation between two points on the sky
-def angsep(ra1,dec1,ra2,dec2):
-	ra1*=d2r
-	dec1*=d2r
-	ra2*=d2r
-	dec2*=d2r
-	diffCosine=cos(dec1)*cos(dec2)*cos(ra1-ra2)+sin(dec1)*sin(dec2)
-	dC='%.10f'%diffCosine#when the source is right at the center of the roi python sometimes adds extraneous digits at the end of the value i.e. instead of 1.0
-	#it returns 1.0000000000000024, which throws an error with the acos function
-	return acos(float(dC))/d2r #returns values between 0 and pi radians
-
-#Check if a given file exists or not
-def fileCheck(file):
-	if (not os.access(file,os.F_OK)):
-		return 0
-	return 1
-
-#copied from Damien's macro
-def parameter_element(free, name, maximum, minimum, scale, value):
-    """Create an XML document parameter description element"""
-    impl = minidom.getDOMImplementation()
-    xmldoc_out = impl.createDocument(None,None,None)
-    parameter = xmldoc_out.createElement('parameter')
-    parameter.setAttribute('free', str(free))
-    parameter.setAttribute('name', str(name))
-    parameter.setAttribute('max', str(maximum))
-    parameter.setAttribute('min', str(minimum))
-    parameter.setAttribute('scale', str(scale))
-    parameter.setAttribute('value', str(value))
-    return parameter
-
 def mybool(Input):
 	return {'True':True,'False':False,'T':True,'F':False,'t':True,'f':False,'TRUE':True,'FALSE':False,"true":True,"false":False,"1":True,"0":False}.get(Input)
 
@@ -731,25 +855,25 @@ def cli():
 	parser.add_argument("-I","--isofile",type=str,default='$(FERMI_DIR)/refdata/fermi/galdiffuse/iso_P8R3_SOURCE_V3_v1.txt',help="Name of isotropic diffuse template for output model, will default to P8R3 SOURCE class model.")
 	parser.add_argument("-i","--isoname",type=str,default='iso_P8R3_SOURCE_V3_v1',help="Name of isotropic diffuse component in output model, default is for P8R3 SOURCE class.")
 	parser.add_argument("-N","--normsonly",type=mybool,default=False,help="Flag to only let the normalizations of parameters be free, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
-	parser.add_argument("-e","--extDir",type=str,default='',help="Path to directory with LAT extended source templates, will default to STs default.")#need to figure out what that is
-	parser.add_argument("-r","--radLim",type=float,default=-1.,help="Radius, in degrees, from ROI center beyond which all source parameters should be fixed, will default to selection radius.")
-	parser.add_argument("-R","--maxRad",type=float,default=None,help="Absolute maximum radius, in degrees, from ROI center beyond which all source parameters should be fixed, even variable sources will not be freed beyond this radius, defaults to radLim value.")
-	parser.add_argument("-ER","--ExtraRad",type=float,default=10.,help="Radius beyond event file ROI out to which sources will be included in the model with all parameters fixed, default is 10, good for analyses starting around a few hundred MeV, can be decreased for high energy only fits.")
-	parser.add_argument("-s","--sigFree",type=float,default=5.,help="Average significance below which all source parameters are fixed, defaults to 5.  Note, if using the 3FGL catalog xml file as input, this is actually a cut on TS, so adjust accordingly.")
-	parser.add_argument("-v","--varFree",type=mybool,default=True,help="Flag to set normalization of significantly variable sources, even if source is beyond radius limit or below TS limit, default is True.",choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
-	parser.add_argument("-p","--psForce",type=mybool,default=False,help="Flag to cast extended sources as point sources, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
-	parser.add_argument("-E2C","--E2CAT",type=mybool,default=False,help="Flag to use catalog names for extended sources, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
-	parser.add_argument("-m","--makeRegion",type=mybool,default=True,help="Flag to create ds9 region file as well as the xml model, default is True.",choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
-	parser.add_argument("-GIF","--GIndexFree",type=mybool,default=False,help="Flag to use a power-law modification to the Galactic diffuse model spectrum and have the index be free, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
+	parser.add_argument("-e","--extended_directory",type=str,default='',help="Path to directory with LAT extended source templates, will default to STs default.")#need to figure out what that is
+	parser.add_argument("-r","--free_radius",type=float,default=-1.,help="Radius, in degrees, from ROI center beyond which all source parameters should be fixed, will default to selection radius.")
+	parser.add_argument("-R","--max_free_radius",type=float,default=None,help="Absolute maximum radius, in degrees, from ROI center beyond which all source parameters should be fixed, even variable sources will not be freed beyond this radius, defaults to free_radius value.")
+	parser.add_argument("-ER","--extra_radius",type=float,default=10.,help="Radius beyond event file ROI out to which sources will be included in the model with all parameters fixed, default is 10, good for analyses starting around a few hundred MeV, can be decreased for high energy only fits.")
+	parser.add_argument("-s","--sigma_to_free",type=float,default=5.,help="Average significance below which all source parameters are fixed, defaults to 5.  Note, if using the 3FGL catalog xml file as input, this is actually a cut on TS, so adjust accordingly.")
+	parser.add_argument("-v","--variable_free",type=mybool,default=True,help="Flag to set normalization of significantly variable sources, even if source is beyond radius limit or below TS limit, default is True.",choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
+	parser.add_argument("-p","--force_point_sources",type=mybool,default=False,help="Flag to cast extended sources as point sources, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
+	parser.add_argument("-E2C","--extended_catalog_names",type=mybool,default=False,help="Flag to use catalog names for extended sources, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
+	parser.add_argument("-m","--make_region",type=mybool,default=True,help="Flag to create ds9 region file as well as the xml model, default is True.",choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
+	parser.add_argument("-GIF","--galactic_index_free",type=mybool,default=False,help="Flag to use a power-law modification to the Galactic diffuse model spectrum and have the index be free, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
 	parser.add_argument("-wd","--writeDir",type=str,default='',help="Directory to write the output ds9 region file in if not the current working directory or if you are specifying the full path to the newly made XML file.")
-	parser.add_argument("-ON","--oldNames",type=mybool,default=False,help="Flag to use the make2FLGxml style naming convention, underscore before name and no spaces, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
+	parser.add_argument("-ON","--use_old_names",type=mybool,default=False,help="Flag to use the make2FLGxml style naming convention, underscore before name and no spaces, default is False.",nargs="?",const=True,choices=['True','False','T','F','t','f','TRUE','FALSE','true','false',1,0])
 	parser.add_argument('-DR','--DataRelease',type=int,default=3,help='Choice of data release, 3 for 4FGL-DR3, 2 for 4FGL-DR2, and 1 for 4FGL.',choices=[1,2,3])
 	
 	args=parser.parse_args()
 
 	sL=srcList(args.catalog,args.ev,args.outputxml,args.DataRelease)
 	
-	sL.makeModel(args.galfile,args.galname,args.isofile,args.isoname,args.normsonly,args.extDir,args.radLim,args.maxRad,args.ExtraRad,args.sigFree,args.varFree,args.psForce,args.E2CAT,args.makeRegion,args.GIndexFree,args.writeDir,args.oldNames)
+	sL.makeModel(args.galfile,args.galname,args.isofile,args.isoname,args.normsonly,args.extended_directory,args.free_radius,args.max_free_radius,args.extra_radius,args.sigma_to_free,args.variable_free,args.force_point_sources,args.extended_catalog_names,args.make_region,args.galactic_index_free,args.writeDir,args.use_old_names)
 	
 	
 if __name__=='__main__': cli()
